update: imports updated for moved files
1. proguard rules for pojo updated
2. activity path in manifest updated
3. imports updated for moved pojos
4. misc imports fixes
