+++ app/src/androidTest/java/com/example/omdb/ExampleInstrumentedTest.kt
@@ -2,11 +2,9 @@ package com.example.omdb
 
 import androidx.test.InstrumentationRegistry
 import androidx.test.runner.AndroidJUnit4
-
+import org.junit.Assert.assertEquals
 import org.junit.Test
 import org.junit.runner.RunWith
-
-import org.junit.Assert.*
 
 /**
  * Instrumented test, which will execute on an Android device.
