+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
         tools:targetApi="s">
 
         <activity
-            android:name=".ui.HomeActivity"
+            android:name=".framework.HomeActivity"
             android:exported="true"
             android:screenOrientation="portrait"
             android:theme="@style/SplashTheme"
