+++ app/src/main/java/com/example/omdb/data/local/dao/EpisodeDao.kt
@@ -4,7 +4,7 @@ import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import com.example.omdb.models.local.EntityEpisode
+import com.example.omdb.data.local.model.EntityEpisode
 
 @Dao
 interface EpisodeDao {
