+++ app/src/main/java/com/example/omdb/data/local/dao/RatingDao.kt
@@ -3,7 +3,7 @@ package com.example.omdb.data.local.dao
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
-import com.example.omdb.models.local.EntityRating
+import com.example.omdb.data.local.model.EntityRating
 
 @Dao
 interface RatingDao {
