+++ app/src/main/java/com/example/omdb/data/local/dao/UserDao.kt
@@ -4,7 +4,7 @@ import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import com.example.omdb.models.local.EntityUser
+import com.example.omdb.data.local.model.EntityUser
 
 @Dao
 interface UserDao {
