+++ app/src/main/java/com/example/omdb/di/ApiModule.kt
@@ -1,12 +1,12 @@
 package com.example.omdb.di
 
-import com.example.omdb.data.api.contract.OMDbApi
-import com.example.omdb.data.api.impl.OMDbApiImpl
+import com.example.omdb.data.network.api.contract.OMDbApi
+import com.example.omdb.data.network.api.impl.OMDbApiImpl
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import io.ktor.client.*
+import io.ktor.client.HttpClient
 import javax.inject.Singleton
 
 @Module
