+++ app/src/main/java/com/example/omdb/util/wrappers/FlowWrapper.kt
@@ -1,6 +1,6 @@
 package com.example.omdb.util.wrappers
 
-import com.example.omdb.models.Resource
+import com.example.omdb.domain.model.Resource
 import com.example.omdb.util.extensions.asIoFlow
 import com.example.omdb.util.extensions.asResourceFlow
 import kotlinx.coroutines.flow.Flow
